[td] log error instead of throwing
In clients whom intentionally decide to not initialise TelemetryDeck (for example on non-produciton environments, unit tests, etc), they would have to add a wrapper around all the interactions in order to avoid a situation which would throw 500s when used in requests.

This wasn't intentional, and the error didn't seem relevant enough to warrant a reaction by the client. If you've not registered - it's probably for a reason.

So we've replaced it with a log instead. May consider making it a one-off, will consider options.